+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Always drop empty values for `intExplode` calls (#1000)
 - Discard invalid gender values (#996)
 - Avoid some version-specific tests (#964, #967)
 - Avoid deprecated `settingLanguage` calls (#955)